Moved send_packet impls to its own mod
