+++ crates/relayer-all-in-one/src/one_for_all/impls/chain/queries/mod.rs
@@ -2,6 +2,7 @@ pub mod channel;
 pub mod consensus_state;
 pub mod packet_commitments;
 pub mod received_packet;
+pub mod send_packet;
 pub mod status;
 pub mod unreceived_packets;
 pub mod write_ack;
